+++ src/include/migraphx/program.hpp
@@ -1,7 +1,7 @@
 /*
  * The MIT License (MIT)
  *
- * Copyright (c) 2015-2025 Advanced Micro Devices, Inc. All rights reserved.
+ * Copyright (c) 2015-2026 Advanced Micro Devices, Inc. All rights reserved.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -56,6 +56,7 @@ struct marker;
  */
 struct MIGRAPHX_EXPORT program
 {
+
     program();
 
     explicit program(module m);
@@ -78,6 +79,8 @@ struct MIGRAPHX_EXPORT program
     instruction_ref get_parameter(std::string name) const;
 
     std::unordered_map<std::string, shape> get_parameter_shapes() const;
+
+    int get_program_file_version() const;
 
     std::size_t total_instructions() const;
 
@@ -165,6 +168,9 @@ struct MIGRAPHX_EXPORT program
     private:
     void assign(const program& p);
     std::unique_ptr<program_impl> impl;
+    // program file version is for the data structure or format of the MXR file. Version should be bumped
+    // if any changes occur to the format of the MXR file.
+    static constexpr int program_file_version = 8;
 };
 } // namespace MIGRAPHX_INLINE_NS
 } // namespace migraphx